Some changes in the logic and bug fixes
I have switched to respecting the groupness of neighbours now too. If
atom i is in the fix group, but atom j is not in the same group then
that atom does not exist for the fix. In the previous fix_eph I include
those atoms even though they do not belong to the group. I have to
discuss with Alfredo about the approach.